Added files so far
Description of each file so far:
main.py:
Has a class called Chess that that contains different functions that initialize the main menu, and the different screens.

screens.py: 
Contains the logic of how things are displayed on each screen, except from popping the pieces on the board for the player screen.
button.py: 
contains a class with the logic for the buttons of the game and the hover functionality.
board.py:
has the logic to initialize the board with the pieces on a matrix of two levels (list of lists), also has the logic to update the board when a piece has been moved on the list that saves the current state of the board.
movements.py:
has the logic for moving the pieces (without any rules yet), functions to select a piece, another one select the desired square, and another one to move the piece updating the list of the board which is taken as an input on the main file. There's also a method to convert the mouse position to row call to match it with the rows columns of the list.