Bliting captured pieces
As part of finishing the GUI before working on the chess logic, added a
method to display on screen the pieces that have been captured on the
current game.
So far the only thing that the method does is to append the captured
piece to a list called, captured pieces, looping through them will be
added in the upcoming days or perhaps to a new method.