En passant
Added the en passant rule for the allowed moves on the pawn function.

Be aware that I programed this to check is the pawn we selected is on
the 3 or 4th row depending on the color of the piece, however I yet
haven't restricted this to only allow en passant when the enemy pawn
advances two squares, it will also allow en passant if the pawn was
already there, this needs to be fixed.