En passant correction
Corrected some bugs for the En passant functionality, now it prevents
for you to capture a pawn that hasn't moved two squares in the turn
immediatly before yours. Also had a bug not letting me capture a piece
on a regular capture if I was on the en passant row, this was corrected.
Currently there's nothign weird to be aware of, will start implementing
rules for other pieces.