Function to define horizontal moves
Corrected some formating issues.

The approach that I'm going to take to define the movements of the
pieces moving forward is to create three functions, one to define
horizontal moves, one to define vertical moves and one to define
diagonal moves, then call the function on each piece as needed. For
obvious reasons the knight will have a separate method for it's own.