Change to horizontal moves
Made a change for to the _define_horizontal_moves() for it to work with
vertical moves too dinamically. While maintaining it's identity of being
able to work for the queen. Not sure if I included this on the previous
commit but this function allows captures too and seems to be working
completely fine witht the function to draw the captures (and movements)
to the screen

Speaking of that func, beaware that it does not handle the chess
notation rules that works differntly if two pieces can capture on one
same square, this will be implemented in the future.