Castle
Made a preliminary implementation of the special movement castle.

Beaware, this only takes care of allowing the king to castle, does not
check if the king is in check (there is no way to check the check yet),
does not check if any of the castling pieces has moved and has a bug
when castling on the turn of the other player.