Check
Added an initial implementation of the function to determine if the king
is in check, not working well but not generating any bugs since it only
prints to the conmsole. the idea is to iterate over all the pieces at
the begining of each turn and determine ig any of those are attacking
the king. This might change in the future. Also while writing this
commit I realized that this approach perhaps is not able to prevent the
king moving to a check postion, this might not be the best approach.