Check function
Fixed the bug that was not identifying when the king was in check adding
a variable which is check_turn (opposite of the current turn) to
properly calculate the check. Appears to be working with all pieces but
still haven't where to use it.

Be aware that currently it's being used at the begining of the
move_piece() and it won't let you move any piece if at some point you
are in check, not even to get out of the check