_out_of_check()
Implemented a new method called out of check that is called mainly to
understand if after the intended move the king is out of check.

Created a deep copy of the board and of the king_pos (updated
accordingly) to manage when the king exits the check by moving

Be aware that at some point this had a bug when calling the function to
calculate the moves (bc this functions calls the check func), but I
haven't seen this presenting again after some changes that I made
working on some other stuff, will debug this in teh upcoming days.