Bug with diagonal moves on check functions
The functions that checked if the king was at check were having a bug
when the king tried to castle when at check. Traced the error down to
the diagonal moves function wich was not succesfully returning when the
list was out of range.

Fixed the condition to return if any of the values were out of range.

Be aware: King is able to get out of the check castling, which is
agains the rules, and im pretty certain that it doesn't prevent castling
if any of teh squares of teh castle would be at check, will be fixing
this. Also, haven't implementd yet the notation for the check.