Checking parameter
Added this checking parameter to the function that defines the valid
moves for it not to execute the castle function when checking for a way
to get out of the check since the castle is not a valid move to get out
of the check. Also moved the conditon that determiens if we are in check
at the very begining of the move piece function so that it passes as an
argument that we are checking true, so that it doesnt execute the
castle.